+++ src/engine/wasmEngine/wasm/src/asc/raycaster/player.ts
@@ -4,76 +4,76 @@ import { ObjectAllocator, newObjectAllocator } from '../objectAllocator';
 import { logi } from '../importVars';
 
 @final @unmanaged class Player {
-  private posX: f64;
-  private posY: f64;
-  private posZ: f64;
-  private dirX: f64;
-  private dirY: f64;
-  private planeX: f64;
-  private planeY: f64;
-  private pitch: f64;
+  private posX: f32;
+  private posY: f32;
+  private posZ: f32;
+  private dirX: f32;
+  private dirY: f32;
+  private planeX: f32;
+  private planeY: f32;
+  private pitch: f32;
 
-  get PosX(): f64 {
+  get PosX(): f32 {
     return this.posX;
   }
   
-  set PosX(posX: f64) {
+  set PosX(posX: f32) {
     this.posX = posX;
   }
 
-  get PosY(): f64 {
+  get PosY(): f32 {
     return this.posY;
   }
 
-  set PosY(posY: f64) {
+  set PosY(posY: f32) {
     this.posY = posY;
   }
 
-  get PosZ(): f64 {
+  get PosZ(): f32 {
     return this.posZ;
   }
 
-  set PosZ(posZ: f64) {
+  set PosZ(posZ: f32) {
     this.posZ = posZ;
   }
 
-  get DirX(): f64 {
+  get DirX(): f32 {
     return this.dirX;
   }
 
-  set DirX(dirX: f64) {
+  set DirX(dirX: f32) {
     this.dirX = dirX;
   }
 
-  get DirY(): f64 {
+  get DirY(): f32 {
     return this.dirY;
   }
 
-  set DirY(dirY: f64) {
+  set DirY(dirY: f32) {
     this.dirY = dirY;
   }
 
-  get PlaneX(): f64 {
+  get PlaneX(): f32 {
     return this.planeX;
   }
 
-  set PlaneX(planeX: f64) {
+  set PlaneX(planeX: f32) {
     this.planeX = planeX;
   }
 
-  get PlaneY(): f64 {
+  get PlaneY(): f32 {
     return this.planeY;
   }
 
-  set PlaneY(planeY: f64) {
+  set PlaneY(planeY: f32) {
     this.planeY = planeY;
   }
 
-  get Pitch(): f64 {
+  get Pitch(): f32 {
     return this.pitch;
   }
 
-  set Pitch(pitch: f64) {
+  set Pitch(pitch: f32) {
     this.pitch = pitch;
   }
 }
